Fix bug where pool wouldn't close due how we close
We close the pool by sending an element into the closeCh for each
thread we have in the pool. However, some threads (Producers) can
close early due to completing their workload. This resulted in a
blocking call when adding any elements to the closeCh terminated
threads.

We changed the closeCh to have a buffer equivalent to the number of
threads in the buffer and only add an element into the closeCh for
threads that have not yet completed. Additionally we added a test
to catch this in the future.